wip
Former-commit-id: 11c55af7f605b7e4d52ce635e92357243b8d190f

+++ apps/dns/internal/domain/main.go
@@ -164,7 +164,7 @@ func (d domainI) CreateRecord(
 	recordType string,
 	host string,
 	answer string,
-	ttl int64,
+	ttl uint32,
 	priority int64,
 ) (*Record, error) {
 	create, err := d.recordsRepo.Create(ctx, &Record{
